backend/pyproject.toml
[project]
name = "deliphone-backend"
version = "0.1.0"
description = "Deliphone — rental service backend (FastAPI)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "Deliphone" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "alembic>=1.14",
    "celery>=5.4",
    "redis>=5.2",
    "structlog>=24.4",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.2",
    "httpx>=0.28",
    "boto3>=1.35",
    "geoalchemy2>=0.15",
    "pyotp>=2.9",
    "cryptography>=43",
    "email-validator>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-python-jose",
    "types-passlib",
    "freezegun>=1.5",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ---------------- Ruff ----------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["B008"]  # FastAPI Depends() in defaults is idiomatic

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["E501"]
"tests/*" = ["S101"]

# ---------------- mypy ----------------
[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ["alembic/versions"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

# ---------------- pytest ----------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
